Split BinarySearch class
