+++ Source/AtCoderLibrary/STL/Extension/BinarySearch.Span.cs
@@ -35,17 +35,6 @@ namespace AtCoder.Extension
         /// </remarks>
         [MethodImpl(256)]
         public static int LowerBound<T>(this T[] a, T v) where T : IComparable<T>
-            => LowerBound(a.AsSpan(), v, default(DefaultComparer<T>));
-
-        /// <summary>
-        /// <paramref name="a"/> の要素のうち、<paramref name="v"/> 以上の値が現れる最小のインデックスを取得します。
-        /// </summary>
-        /// <remarks>
-        /// <para>制約: <paramref name="a"/> がソート済み</para>
-        /// <para>計算量: O(n)</para>
-        /// </remarks>
-        [MethodImpl(256)]
-        public static int LowerBound<T>(this IList<T> a, T v) where T : IComparable<T>
             => LowerBound(a, v, default(DefaultComparer<T>));
 
         /// <summary>
@@ -89,17 +78,6 @@ namespace AtCoder.Extension
         /// <para>計算量: O(n)</para>
         /// </remarks>
         [MethodImpl(256)]
-        public static int LowerBound<T, TOp>(this IList<T> a, T v, TOp cmp) where TOp : IComparer<T>
-            => BinarySearch<T, TOp, L>(a, v, cmp);
-
-        /// <summary>
-        /// <paramref name="a"/> の要素のうち、<paramref name="v"/> 以上の値が現れる最小のインデックスを取得します。
-        /// </summary>
-        /// <remarks>
-        /// <para>制約: <paramref name="a"/> がソート済み</para>
-        /// <para>計算量: O(n)</para>
-        /// </remarks>
-        [MethodImpl(256)]
         public static int LowerBound<T, TOp>(this Span<T> a, T v, TOp cmp) where TOp : IComparer<T>
             => BinarySearch<T, TOp, L>(a, v, cmp);
 
@@ -124,17 +102,6 @@ namespace AtCoder.Extension
         [MethodImpl(256)]
         public static int UpperBound<T>(this T[] a, T v) where T : IComparable<T>
             => UpperBound(a.AsSpan(), v, default(DefaultComparer<T>));
-
-        /// <summary>
-        /// <paramref name="a"/> の要素のうち、<paramref name="v"/> より大きい値が現れる最小のインデックスを取得します。
-        /// </summary>
-        /// <remarks>
-        /// <para>制約: <paramref name="a"/> がソート済み</para>
-        /// <para>計算量: O(n)</para>
-        /// </remarks>
-        [MethodImpl(256)]
-        public static int UpperBound<T>(this IList<T> a, T v) where T : IComparable<T>
-            => UpperBound(a, v, default(DefaultComparer<T>));
 
         /// <summary>
         /// <paramref name="a"/> の要素のうち、<paramref name="v"/> より大きい値が現れる最小のインデックスを取得します。
@@ -177,17 +144,6 @@ namespace AtCoder.Extension
         /// <para>計算量: O(n)</para>
         /// </remarks>
         [MethodImpl(256)]
-        public static int UpperBound<T, TOp>(this IList<T> a, T v, TOp cmp) where TOp : IComparer<T>
-            => BinarySearch<T, TOp, U>(a, v, cmp);
-
-        /// <summary>
-        /// <paramref name="a"/> の要素のうち、<paramref name="v"/> より大きい値が現れる最小のインデックスを取得します。
-        /// </summary>
-        /// <remarks>
-        /// <para>制約: <paramref name="a"/> がソート済み</para>
-        /// <para>計算量: O(n)</para>
-        /// </remarks>
-        [MethodImpl(256)]
         public static int UpperBound<T, TOp>(this Span<T> a, T v, TOp cmp) where TOp : IComparer<T>
             => BinarySearch<T, TOp, U>(a, v, cmp);
 
@@ -202,22 +158,6 @@ namespace AtCoder.Extension
         public static int UpperBound<T, TOp>(this ReadOnlySpan<T> a, T v, TOp cmp) where TOp : IComparer<T>
             => BinarySearch<T, TOp, U>(a, v, cmp);
 
-        [MethodImpl(256)]
-        private static int BinarySearch<T, TOp, TOk>(this IList<T> a, T v, TOp cmp)
-            where TOp : IComparer<T>
-            where TOk : IOk
-        {
-            int ok = a.Count;
-            int ng = -1;
-            while (ok - ng > 1)
-            {
-                var m = (ok + ng) >> 1;
-                var c = cmp.Compare(a[m], v);
-                if (default(TOk).Ok(c)) ok = m;
-                else ng = m;
-            }
-            return ok;
-        }
         [MethodImpl(256)]
         private static int BinarySearch<T, TOp, TOk>(this ReadOnlySpan<T> a, T v, TOp cmp)
             where TOp : IComparer<T>
